+++ web_src/src/components/AgentSidebar/useApplyAiProposal.ts
@@ -1,5 +1,4 @@
 import type { CanvasChangesetChange } from "@/api-client";
-import type { Dispatch, SetStateAction } from "react";
 import { useCallback } from "react";
 import type { AiBuilderMessage, AiBuilderProposal } from "./agentChat";
 import { pushAiMessages } from "./agentChat";
@@ -7,10 +6,10 @@ import { pushAiMessages } from "./agentChat";
 export type UseApplyAiProposalParams = {
   onApplyAiOperations?: (changes: CanvasChangesetChange[]) => Promise<void>;
   pendingProposal: AiBuilderProposal | null;
-  setAiError: Dispatch<SetStateAction<string | null>>;
-  setIsApplyingProposal: Dispatch<SetStateAction<boolean>>;
-  setAiMessages: Dispatch<SetStateAction<AiBuilderMessage[]>>;
-  setPendingProposal: Dispatch<SetStateAction<AiBuilderProposal | null>>;
+  setAiError: (error: string | null) => void;
+  setIsApplyingProposal: (isApplying: boolean) => void;
+  setAiMessages: (messages: AiBuilderMessage[]) => void;
+  setPendingProposal: (proposal: AiBuilderProposal | null) => void;
 };
 
 export function useApplyAiProposal({
